requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
python-jose
passlib[bcrypt]
bcrypt==3.2.2
pydantic[email]
python-multipart